refactor: make SSR business logic agnostic
