+++ src/client/Wrapper.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import url from 'url';
 import { connect } from 'react-redux';
 import { IntlProvider, addLocaleData } from 'react-intl';
 import { withRouter } from 'react-router-dom';
@@ -13,6 +14,7 @@ import {
   getAuthenticatedUserName,
   getLocale,
   getUsedLocale,
+  getUseBeta,
 } from './reducers';
 import { login, logout, busyLogin } from './auth/authActions';
 import { getFollowing, getNotifications } from './user/userActions';
@@ -22,6 +24,7 @@ import {
   getTrendingTopics,
   setUsedLocale,
   busyAPIHandler,
+  setAppUrl,
 } from './app/appActions';
 import * as reblogActions from './app/Reblog/reblogActions';
 import busyAPI from './busyAPI';
@@ -90,16 +93,47 @@ export default class Wrapper extends React.PureComponent {
     busyAPIHandler: () => {},
   };
 
-  static fetchData(store) {
-    return store.dispatch(login());
+  static async fetchData({ store, req, res }) {
+    await store.dispatch(login());
+
+    const appUrl = url.format({
+      protocol: req.protocol,
+      host: req.get('host'),
+    });
+
+    store.dispatch(setAppUrl(appUrl));
+
+    const state = store.getState();
+
+    const useBeta = getUseBeta(state);
+
+    if (useBeta && appUrl === 'https://busy.org') {
+      res.redirect(`https://staging.busy.org${req.originalUrl}`);
+      return;
+    }
+
+    const locale = getLocale(state);
+
+    await Wrapper.loadLocaleData(locale);
+
+    store.dispatch(setUsedLocale(getAvailableLocale(locale)));
+  }
+
+  static async loadLocaleData(locale) {
+    const availableLocale = getAvailableLocale(locale);
+    const translationsLocale = getTranslationsByLocale(locale);
+
+    const localeDataPromise = import(`react-intl/locale-data/${availableLocale}`);
+    const translationsPromise = import(`./locales/${translationsLocale}.json`);
+
+    const [localeData, translations] = await Promise.all([localeDataPromise, translationsPromise]);
+
+    addLocaleData(localeData);
+    global.translations = translations;
   }
 
   constructor(props) {
     super(props);
-
-    this.state = {
-      translations: global.translations,
-    };
 
     this.loadLocale = this.loadLocale.bind(this);
     this.handleMenuItemClick = this.handleMenuItemClick.bind(this);
@@ -136,22 +170,9 @@ export default class Wrapper extends React.PureComponent {
     }
   }
 
-  loadLocale(locale) {
-    const availableLocale = getAvailableLocale(locale);
-    const translationsLocale = getTranslationsByLocale(locale);
-
-    const localeDataPromise = import(`react-intl/locale-data/${availableLocale}`);
-    const translationsPromise = import(`./locales/${translationsLocale}.json`);
-
-    Promise.all([localeDataPromise, translationsPromise]).then(([localeData, translations]) => {
-      addLocaleData(localeData);
-      this.setState(
-        {
-          translations,
-        },
-        () => this.props.setUsedLocale(availableLocale),
-      );
-    });
+  async loadLocale(locale) {
+    await Wrapper.loadLocaleData(locale);
+    this.props.setUsedLocale(getAvailableLocale(locale));
   }
 
   handleMenuItemClick(key) {
@@ -193,10 +214,9 @@ export default class Wrapper extends React.PureComponent {
 
   render() {
     const { user, usedLocale, locale } = this.props;
-    const { translations } = this.state;
 
     return (
-      <IntlProvider key={usedLocale} locale={usedLocale} messages={translations}>
+      <IntlProvider key={usedLocale} locale={usedLocale} messages={global.translations}>
         <LocaleProvider locale={enUS}>
           <Layout data-dir={getLocaleDirection(getAvailableLocale(locale))}>
             <Layout.Header style={{ position: 'fixed', width: '100vw', zIndex: 1050 }}>
